pyproject.toml
[tool.poetry]
name = "colour-workbench"
version = "0.2.0a"
description = "A workbench environment for developing colour algorithms."
authors = ["Tucker Downs <tucker@tjdcs.dev>"]
maintainers = ["Tucker Downs <tucker@tjdcs.dev>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/tjdcs/colour-workbench"
packages = [{ include = "colour_workbench" }]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
colour-science = { path = "submodules/colour-science", develop = true }
colour-datasets = { path = "submodules/colour-datasets", develop = true }
colour-specio = { path = "submodules/colour-specio", develop = true }
numpy = "^1.25.0"
scipy = "^1.10.1"
xxhash = "^3.2.0"
matplotlib = "^3.7.1"
scikit-learn = "^1.2.2"
requests = "^2.31.0"
python-tsp = "^0.3.1"
pandas = "^2.0.3"
shapely = "^2.0.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^23.3.0"
pytest = "^7.4.0"
pytest-xdist = "^3.3.1"
pyright = "^1.1.332"
ruff = "^0.1.0"
flynt = "^1.0.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Begin Colour-Style
[tool.black]
line-length = 79
exclude = '''
/(
    \.git
  | build
  | dist
)/
'''

[tool.flynt]
line_length = 999

[tool.pyright]
reportMissingImports = false
reportMissingModuleSource = false
reportUnboundVariable = false
reportUnnecessaryCast = true
reportUnnecessaryTypeIgnoreComment = true
reportUnsupportedDunderAll = false
reportUnusedExpression = false

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
split_on_trailing_comma = true
skip_glob = ["colour/**/__init__.py"]

[tool.ruff]
target-version = "py39"
line-length = 88
ignore-init-module-imports = true
select = [
    "A",   # flake8-builtins
    "ARG", # flake8-unused-arguments
    # "ANN", # flake8-annotations
    "B", # flake8-bugbear
    # "BLE", # flake8-blind-except
    "C4", # flake8-comprehensions
    # "C90", # mccabe
    # "COM", # flake8-commas
    "DTZ", # flake8-datetimez
    "D",   # pydocstyle
    "E",   # pydocstyle
    # "ERA", # eradicate
    # "EM", # flake8-errmsg
    "EXE", # flake8-executable
    "F",   # flake8
    # "FBT", # flake8-boolean-trap
    "G",   # flake8-logging-format
    "I",   # isort
    "ICN", # flake8-import-conventions
    "INP", # flake8-no-pep420
    "ISC", # flake8-implicit-str-concat
    "N",   # pep8-naming
    # "PD", # pandas-vet
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "PL",  # pylint
    # "PT", # flake8-pytest-style
    # "PTH", # flake8-use-pathlib [Enable]    "Q",   # flake8-quotes
    "RET", # flake8-return
    "RUF", # Ruff
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    # "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "TRY", # tryceratops
    "UP",  # pyupgrade
    "W",   # pydocstyle
    "YTT", # flake8-2020
]
ignore = [
    "B008",
    "B905",
    "D104",
    "D200",
    "D202",
    "D205",
    "D301",
    "D400",
    "I001",
    "N801",
    "N802",
    "N803",
    "N806",
    "N813",
    "N815",
    "N816",
    "ICN001",
    "PGH003",
    "PIE804",
    "PLE0605",
    "PLR0911",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "RET508",
    "TRY003",
    "TRY300",
]

fixable = ["B", "C", "E", "F", "PIE", "RUF", "SIM", "UP", "W"]

[tool.ruff.pydocstyle]
convention = "numpy"
# End Colour-Style
